requirements.txt
biopython>=1.83
python-docx>=1.1.0
google-generativeai>=0.8.0
google-genai>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
